refactored and seperated main.py
