+++ .gitignore
@@ -1,0 +1,3 @@
+testing_notebook.ipynb
+__pycache__
+.pytest_cache
