+++ test/picrosstest.py
@@ -1,0 +1,19 @@
+from numpy import array_equal
+from src.solver import solve_puzzle 
+
+def test_picross():
+    clues = [[[2],[1,1],[4],[2,1],[3,1],[8],[8],[7],[5],[3]],[[1],[2],[1,6],[9],[6],[5],[5],[4],[3],[4]]]
+    pix_grid = solve_puzzle(clues)
+
+    assert array_equal(pix_grid, 
+                       [[-1., -1.,  1.,  1., -1., -1., -1., -1., -1., -1.],
+                        [-1.,  1., -1.,  1., -1., -1., -1., -1., -1., -1.],
+                        [ 1.,  1.,  1.,  1., -1., -1., -1., -1., -1., -1.],
+                        [-1., -1.,  1.,  1., -1., -1., -1., -1., -1.,  1.],
+                        [-1., -1.,  1.,  1.,  1., -1., -1., -1., -1.,  1.],
+                        [-1., -1.,  1.,  1.,  1.,  1.,  1.,  1.,  1.,  1.],
+                        [-1., -1.,  1.,  1.,  1.,  1.,  1.,  1.,  1.,  1.],
+                        [-1., -1.,  1.,  1.,  1.,  1.,  1.,  1.,  1., -1.],
+                        [-1., -1., -1.,  1.,  1.,  1.,  1.,  1., -1., -1.],
+                        [-1., -1., -1., -1.,  1.,  1.,  1., -1., -1., -1.]]
+    )
